# server/openapi.yaml
openapi: 3.0.3
info:
  title: Parc Fermé Store API
  version: 1.0.0
  description: >-
    REST API for the Parc Fermé demo motorsport-memorabilia store. All responses include an `X-Request-Id` header for
    tracing. Errors use a consistent envelope: `{ "error": { "code", "message", "details?", "requestId" } }`. **Demo
    accounts** — admin: `admin@parcferme.dev` / `Admin123!`, customer: `ava@demo.dev` / `Customer123!`. **Test cards** —
    `4242424242424242` succeeds, `4000000000000002` declines, `4000000000009995` insufficient funds.
servers:
  - url: /
    description: This host
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details: {}
            requestId:
              type: string
              format: uuid
...
    Order:
      type: object
      properties:
        id:
          type: string
          example: PF-1A7F2K
        status:
          type: string
          enum:
            - pending_payment
            - paid
            - shipped
            - delivered
            - cancelled
        subtotalCents:
          type: integer
        shippingCents:
          type: integer
        taxCents:
          type: integer
        totalCents:
          type: integer
        shipping:
          type: object
        items:
          type: array
          items:
            type: object
        createdAt:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            email:
              type: string
            name:
              type: string
            role:
              type: string
              enum:
                - customer
                - admin
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: Ava Chen
                email:
                  type: string
                  example: new@demo.dev
                password:
                  type: string
                  example: Password123!
      responses:
        '201':
          description: 'Account created (X-Auth-Event: register-success)'
          headers:
            X-Auth-Event:
              description: Auth outcome signal consumed by Fastly NGWAF templated rules
              schema:
                type: string
                enum:
                  - login-success
                  - login-failure
                  - register-success
                  - register-failure
                  - password-reset-attempt
                  - password-reset-success
                  - password-reset-failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 'Validation error (X-Auth-Event: register-failure)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 'Email already registered (X-Auth-Event: register-failure)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Log in and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: ava@demo.dev
                password:
                  type: string
                  example: Customer123!
      responses:
        '200':
          description: 'Authenticated (X-Auth-Event: login-success)'
          headers:
            X-Auth-Event:
              description: Auth outcome signal consumed by Fastly NGWAF templated rules
              schema:
                type: string
                enum:
                  - login-success
                  - login-failure
                  - register-success
                  - register-failure
                  - password-reset-attempt
                  - password-reset-success
                  - password-reset-failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 'Invalid credentials (X-Auth-Event: login-failure)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request a password reset link
      description: >-
        Always returns 200 with an identical body (anti-enumeration). Sets response header `X-Auth-Event:
        password-reset-attempt`. No email is sent in this demo; the reset link is logged server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  example: ava@demo.dev
      responses:
        '200':
          description: 'Reset requested (X-Auth-Event: password-reset-attempt)'
          headers:
            X-Auth-Event:
              description: Auth outcome signal consumed by Fastly NGWAF templated rules
              schema:
                type: string
                enum:
                  - login-success
                  - login-failure
                  - register-success
                  - register-failure
                  - password-reset-attempt
                  - password-reset-success
                  - password-reset-failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: 'Validation error (X-Auth-Event: password-reset-failure)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset a password using a reset token
      description: >-
        Sets `X-Auth-Event: password-reset-success` on success, or `password-reset-failure` on any bad/expired/used
        token or invalid new password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  example: a1b2c3…
                password:
                  type: string
                  example: NewPassword123!
      responses:
        '200':
          description: 'Password reset (X-Auth-Event: password-reset-success)'
          headers:
            X-Auth-Event:
              description: Auth outcome signal consumed by Fastly NGWAF templated rules
              schema:
                type: string
                enum:
                  - login-success
                  - login-failure
                  - register-success
                  - register-failure
                  - password-reset-attempt
                  - password-reset-success
                  - password-reset-failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: 'Invalid/expired token or weak password (X-Auth-Event: password-reset-failure)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
        '401':
          description: Missing/invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/orders:
    post:
      tags:
        - Orders
      summary: Create an order from a cart (requires auth)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cartId
                - shipping
              properties:
                cartId:
                  type: string
                  format: uuid
                shipping:
                  type: object
                  required:
                    - name
                    - line1
                    - city
                    - postalCode
                    - country
                  properties:
                    name:
                      type: string
                    line1:
                      type: string
                    line2:
                      type: string
                    city:
                      type: string
                    postalCode:
                      type: string
                    country:
                      type: string
      responses:
        '201':
          description: Order created (pending_payment)
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Empty cart / stock issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Auth required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Orders
      summary: List my orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order history
        '401':
          description: Auth required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/orders:
    get:
      tags:
        - Admin
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending_payment
              - paid
              - shipped
              - delivered
              - cancelled
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Orders
...